# LibJS: Make local variables faster by avoiding the environment when possible