--- a/Userland/Libraries/LibJS/AST.cpp
+++ b/Userland/Libraries/LibJS/AST.cpp
@@ -2679,7 +2679,11 @@ Completion Identifier::execute(Interpreter& interpreter) const
 void Identifier::dump(int indent) const
 {
     print_indent(indent);
-    outln("Identifier \"{}\"", m_string);
+    if (is_local()) {
+        outln("Identifier \"{}\" is_local=(true) index=({})", m_string, m_local_variable_index);
+    } else {
+        outln("Identifier \"{}\"", m_string);
+    }
 }
 
 Completion PrivateIdentifier::execute(Interpreter&) const
@@ -4429,6 +4433,16 @@ ThrowCompletionOr<void> ScopeNode::for_each_var_function_declaration_in_reverse_
 {
     for (ssize_t i = m_var_declarations.size() - 1; i >= 0; i--) {
         auto& declaration = m_var_declarations[i];
+        if (is<FunctionDeclaration>(declaration))
+            TRY(callback(static_cast<FunctionDeclaration const&>(*declaration)));
+    }
+    return {};
+}
+
+ThrowCompletionOr<void> ScopeNode::for_each_lexical_function_declaration_in_reverse_order(ThrowCompletionOrVoidCallback<FunctionDeclaration const&>&& callback) const
+{
+    for (ssize_t i = m_lexical_declarations.size() - 1; i >= 0; i--) {
+        auto& declaration = m_lexical_declarations[i];
         if (is<FunctionDeclaration>(declaration))
             TRY(callback(static_cast<FunctionDeclaration const&>(*declaration)));
     }