… Userland/Libraries/LibJS/Bytecode/Generator.cpp
… Userland/Libraries/LibJS/Bytecode/Generator.h
… Userland/Libraries/LibJS/Bytecode/Instruction.h
… Userland/Libraries/LibJS/Bytecode/Op.cpp
… Userland/Libraries/LibJS/Bytecode/Op.h
… Userland/Libraries/LibJS/Forward.h
… Userland/Libraries/LibJS/Parser.cpp
… Userland/Libraries/LibJS/Parser.h
… Userland/Libraries/LibJS/Runtime/AbstractOperations.cpp
… Userland/Libraries/LibJS/Runtime/ECMAScriptFunctionObject.cpp
… Userland/Libraries/LibJS/Runtime/ECMAScriptFunctionObject.h
… Userland/Libraries/LibJS/Runtime/ExecutionContext.cpp
… Userland/Libraries/LibJS/Runtime/ExecutionContext.h
… Userland/Libraries/LibJS/Runtime/FunctionConstructor.cpp
… Userland/Libraries/LibJS/Runtime/FunctionObject.h
… Userland/Libraries/LibJS/SourceTextModule.cpp
… Userland/Libraries/LibWeb/DOM/EventTarget.cpp
… Userland/Libraries/LibWeb/WebDriver/ExecuteScript.cpp