Implements https://github.com/SerenityOS/serenity/issues/17968

Kraken tests:

ai-astar.js
before: 16.64s after: 12.23s

stanford-crypto-aes.js
before: 3.81s, after: 2.93s

stanford-crypto-pbkdf2.js
before: 7.69s after: 5.48s