@Hendiadyoin1 2023-07-02
Can you check the `EliminateLoads` pass of the optimizer, if anything needs to change there, to accommodate your changes