@awesomekling 2023-07-05
Really nice Kraken progression from this, wow :)