@awesomekling 2023-07-05
I think locals and bytecode registers can be the same thing eventually, but we can move towards that incrementally :)